+++ components/HorizontalScrollFeature.vue
@@ -1,18 +1,4 @@
-<script setup>
-import { ref, computed, onMounted } from 'vue'
-
-const props = defineProps({
-  //   propVar: {
-  //     type: Boolean,
-  //     default: false,
-  //   },
-})
-
-//const emit = defineEmits(["change", "click"]);
-
-// lifecycle hooks
-onMounted(() => {})
-</script>
+<script setup></script>
 
 <template>
   <div class="horizontal-scroll-feature">
