+++ pages/home.vue
@@ -54,5 +54,3 @@ useHead({
     <WNYCFeatured />
   </div>
 </template>
-
-<style lang="scss" scoped></style>
